Front end Ok

+++ database/migrations/2023_01_04_085735_create_produits_table.php
@@ -20,10 +20,10 @@ return new class extends Migration
             $table->foreignId('couleur_id')->constrained();
             $table->string('image')->nullable();
             $table->string('imageFile')->nullable();
-            $table->integer('prix');
-            $table->integer('promo')->default(0);
-            $table->integer('prixfinal');
-            $table->integer('stock');
+            $table->double('prix');
+            $table->double('promo')->default(0);
+            $table->double('prixfinal');
+            $table->double('stock');
             $table->text('description');
             $table->timestamps();
         });
